requirements.txt
Flask==3.0.0
groq==0.4.0
pdfminer.six==20221105
python-dotenv==1.0.0
gunicorn==21.2.0
